separate file for setting up middleware
